Map the aspect ratio from width/height to CSS
Instead of having special code in image layout, map the aspect ratio that's
computed from the width and height attributes to the aspect-ratio property.

As a side-effect, this patch will also make the aspect-ratio property work
when contain: size is specified in an image (as well as error images when
no alt text is specified).

As specified in https://github.com/whatwg/html/pull/6032.

R=masonfreed@chromium.org, xiaochengh@chromium.org

Bug: 1137004
Change-Id: I1253fa2ecf1e176711d3c4d7bd4159f466656ae8